Move resizeable columns jquery code into service
